requirements.txt
django
python-dotenv
psycopg2
Pillow
pyotp
qrcode
pyjwt[crypto]
bcrypt
python-dateutil
whitenoise
django-cors-headers
requests
Django
psycopg2-binary
gunicorn
werkzeug
django-extensions
pyOpenSSL
web3
py-solc-x